Edge_detection: miglioramento bordi, blur (gaus., median...), thresholds

+++ source/edge_detection.py
@@ -1,4 +1,5 @@
 from source import globals
+
 import numpy as np
 import cv2
 import matplotlib.pyplot as plt
@@ -6,6 +7,7 @@ import matplotlib.pyplot as plt
 conf = globals.conf
 
 def quantize(frame, nbin):
+    """
     histogram = np.zeros([256], np.float64)
     for h in range(frame.shape[0]):
         for w in range(frame.shape[1]):
@@ -24,16 +26,65 @@ def quantize(frame, nbin):
             ret[h, w] = bins[p // bin_size]
 
     return ret
+    """
+    #Uniform quatization
+    #return np.round(frame*(nbin/255))*(255/nbin)
 
-def edge_detection(frame, debug = False, corners = False):
-    print("Edge detection function.")
+    # USING KMEANS
+
+    # convert the image from the RGB color space to the L*a*b*
+    # color space -- since we will be clustering using k-means
+    # which is based on the euclidean distance, we'll use the
+    # L*a*b* color space where the euclidean distance implies
+    # perceptual meaning
+
+    #image = cv2.cvtColor(frame, cv2.COLOR_BGR2LAB)
+
+    img = frame#image
+    Z = img.reshape((-1, 3))
+
+    # convert to np.float32
+    Z = np.float32(Z)
+
+    # define criteria, number of clusters(K) and apply kmeans()
+    criteria = (cv2.TERM_CRITERIA_EPS + cv2.TERM_CRITERIA_MAX_ITER, 10, 1.0)
+    K = nbin
+    ret, label, center = cv2.kmeans(Z, K, None, criteria, 10, cv2.KMEANS_RANDOM_CENTERS)
+
+    # Now convert back into uint8, and make original image
+    center = np.uint8(center)
+    res = center[label.flatten()]
+    res2 = res.reshape(img.shape)
+
+    #res2 = cv2.cvtColor(res2, cv2.COLOR_LAB2BGR)
+    
+    return res2
+
+
+def edge_detection(frame, debug = False, corners = False, frame_number = 0):
+
+    selected_frame = 0
+    blur = frame
 
     #Gaussian filter
-    blur = cv2.GaussianBlur(frame, (7, 7), 0)
-    #Median
-    #dst = cv2.medianBlur(frame, 13)
+    blur = cv2.GaussianBlur(blur, (9, 9), 0)
+
+    """
+        Median -- UTILE PER LE STATUE?
+        
+        Sembra che funzioni meglio sul riconoscimento degli edge delle statue.
+    """
+    #blur = cv2.medianBlur(blur, 11)
+
     #Bilateral
-    #dst = cv2.bilateralFilter(frame, 5, 5, 20)
+    #blur = cv2.bilateralFilter(blur, 5, 5, 20)
+
+
+
+    if debug and selected_frame == frame_number:
+        cv2.imshow('Blurred image', blur)
+        cv2.waitKey()
+        cv2.destroyAllWindows()
 
     #Try threshold to eliminate white
     gray = cv2.cvtColor(blur, cv2.COLOR_BGR2GRAY)
@@ -42,16 +93,17 @@ def edge_detection(frame, debug = False, corners = False):
 
     #th3 = cv2.medianBlur(th3, 5)
 
-    ret, th2 = cv2.threshold(gray, 0, 200, cv2.THRESH_BINARY + cv2.THRESH_OTSU)
+    #ret, th2 = cv2.threshold(gray, 0, 200, cv2.THRESH_BINARY + cv2.THRESH_OTSU)
     #thresholded_image = frame
-    gray[gray < ret] = 0
+    #gray[gray < ret] = 0
 
     #Quantize
-    q_frame = quantize(gray, 4)
-    if debug:
+    #q_frame = quantize(gray, 4)
+    #q_frame = quantize(blur, 3)
+    if debug and selected_frame == frame_number:
         # cv2.imshow('Threshold', th3)
         cv2.imshow('Gray', gray)
-        cv2.imshow('Quantized', q_frame)
+        #cv2.imshow('Quantized', q_frame)
         cv2.waitKey()
         cv2.destroyAllWindows()
 
@@ -60,39 +112,43 @@ def edge_detection(frame, debug = False, corners = False):
     #thresholded_image[gray > ret] = [0,0,0]
     #thresholded_image[gray > 60] = [0,0,0]
 
-    thresholded_image = q_frame
-
+    #thresholded_image = q_frame
+    """
+    ret = 0
     if debug:
-        cv2.imshow('Blurred image', blur)
-        #cv2.imshow('Thresholded image - Value: {}'.format(ret), thresholded_image)
+        cv2.imshow('Thresholded image - Value: {}'.format(ret), thresholded_image)
         cv2.waitKey()
         cv2.destroyAllWindows()
-    #Canny
+    """
 
-    factor = 7
-    th = 100
-    TH = th * factor
+    #dst = thresholded_image
 
-    # """
-    v = np.median(thresholded_image.astype(np.uint8))
-    # apply automatic Canny edge detection using the computed median
-    th = int(max(0, (1.0 - 0.30) * v))
-    TH = int(min(255, (1.0 + 0.30) * v))
+    dst = blur
 
-    dst = thresholded_image
-
-    #with aperture 5
-    #th = 500
-    #TH = 700
-    # """
 
     #with aperture 3
-    th = 100
-    TH = 200
-    mod_f = cv2.Canny(dst.astype(np.uint8), th, TH, apertureSize=3, L2gradient=True)
+    # thresholds with K-MEANS
+    #th = 175
+    #TH = 196
 
-    if debug:
+    # Canny
+    # Threshold ottimali per aperture 5
+    # th = 550
+    # TH = 1750
+
+    # threshold con un po' più di rumore
+    th = 500
+    TH = 1300
+
+    mod_f = cv2.Canny(dst.astype(np.uint8), th, TH, apertureSize=5, L2gradient=True)
+
+    vis = frame.copy()
+    vis = np.uint8(vis / 2.)
+    vis[mod_f != 0] = (0, 255, 0)
+
+    if debug and selected_frame == frame_number:
        cv2.imshow('Edge detection image', mod_f)
+       cv2.imshow('Edge detection image', vis)
        cv2.waitKey()
        cv2.destroyAllWindows()
 
@@ -117,11 +173,12 @@ def edge_detection(frame, debug = False, corners = False):
         # Threshold between two values
         mod_f2[ np.where((0.1* dst.max() <  dst) & (dst< 0.15 * dst.max()))] = [0, 0, 255]
 
-        if debug:
+        if debug and selected_frame == frame_number:
             cv2.imshow('Corner detection image', mod_f2)
             cv2.waitKey()
             cv2.destroyAllWindows()
 
-    print("Fine edge_detection.")
 
-    return mod_f2.astype(np.uint8)
+
+    #return mod_f2.astype(np.uint8)
+    return vis
